+++ lib/pmdtester/cmd.rb
@@ -7,6 +7,17 @@ module PmdTester
   class Cmd
     extend PmdTester
     def self.execute(cmd)
+      stdout, _stderr, _status = internal_execute(cmd)
+      stdout&.chomp!
+      stdout
+    end
+
+    def self.stderr_of(cmd)
+      _stdout, stderr, _status = internal_execute(cmd)
+      stderr
+    end
+
+    def self.internal_execute(cmd)
       logger.debug "execute command '#{cmd}'"
 
       stdout, stderr, status = Open3.capture3("#{cmd};")
@@ -20,8 +31,10 @@ module PmdTester
 
       stdout&.chomp!
 
-      stdout
+      [stdout, stderr, status]
     end
+
+    private_class_method :internal_execute
   end
 
   # The exception should be raised when the shell command failed.