+++ test/test_pmd_branch_detail.rb
@@ -19,8 +19,10 @@ class TestPmdBranchDetail < Test::Unit::TestCase
     assert_equal('test message', details.branch_last_message)
     assert_equal('test sha', details.branch_last_sha)
     assert_equal('test time', details.execution_time)
-    assert_equal(PmdTester::Cmd.execute('java -version'), details.jdk_version)
-    assert_equal(PmdTester::Cmd.execute('echo $LANG'), details.language)
+    assert_equal(PmdTester::Cmd.stderr_of('java -version'), details.jdk_version)
+    assert_not_empty(details.jdk_version)
+    assert_equal(ENV['LANG'], details.language)
+    assert_not_empty(details.language)
   end
 
   def test_get_path